+++ src/experiments/naive_incremental_compute_tsv.cpp
@@ -12,8 +12,8 @@
 void naive_incremental_compute_tsv(void(*compute)(Digraph*),
                                    Digraph *g,
                                    raw_edge_array &updates,
-                                   const std::vector<unsigned long> &chunks_start_lines,
-                                   unsigned long total_lines) {
+                                   const std::vector<unsigned int> &chunks_start_lines,
+                                   unsigned int total_lines) {
 
     std::cout << "STARTING NAIVE INCREMENTAL COMPUTATION EXPERIMENT..." << std::endl;
 
@@ -22,10 +22,10 @@ void naive_incremental_compute_tsv(void(*compute)(Digraph*),
     fs.open("measurements.csv");
     fs << "Order" << " , " << "Size" << " , " << "Ingestion Rate" << " , " << "Ingestion CPU Time" << " , " << "Computation CPU time" << std::endl;
 
-    unsigned long num_of_chunks = chunks_start_lines.size();
+    unsigned int num_of_chunks = chunks_start_lines.size();
 
-    unsigned long start_line = 0;
-    unsigned long end_line = 0;
+    unsigned int start_line = 0;
+    unsigned int end_line = 0;
 
     for (int i = 0; i < num_of_chunks; i++) {
         start_line = chunks_start_lines[i];
@@ -41,14 +41,14 @@ void naive_incremental_compute_tsv(void(*compute)(Digraph*),
         // Update the graph
         clock_t cpu_begin_update = clock();
 
-        for(unsigned long j = start_line; j<end_line; j++) {
+        for(unsigned int j = start_line; j<end_line; j++) {
             g->add_edge(updates[j][0], updates[j][1]);
         }
 
         clock_t cpu_end_update = clock();
-        double cpu_time_update = double(cpu_end_update - cpu_begin_update) / CLOCKS_PER_SEC;
+        float cpu_time_update = float(cpu_end_update - cpu_begin_update) / CLOCKS_PER_SEC;
 
-        double ingestion_rate = (end_line-start_line) / cpu_time_update;
+        float ingestion_rate = (end_line-start_line) / cpu_time_update;
 
         std::cout << "Finished updating from chunk " << i+1 << std::endl;
         std::cout << "UPDATE TIME FOR chunk " << std::fixed << i+1 << ": " << cpu_time_update << std::endl;
@@ -67,7 +67,7 @@ void naive_incremental_compute_tsv(void(*compute)(Digraph*),
         compute(g);
 
         clock_t cpu_end_compute = clock();
-        double cpu_time_compute = double(cpu_end_compute - cpu_begin_compute) / CLOCKS_PER_SEC;
+        float cpu_time_compute = float(cpu_end_compute - cpu_begin_compute) / CLOCKS_PER_SEC;
 
         std::cout << "Finished computations." << std::endl;
         std::cout << "COMPUTE TIME AFTER chunk " << std::fixed << i+1 << ": " << cpu_time_compute << std::endl;
