+++ src/preprocessor/edge_array_to_graph.h
@@ -4,7 +4,7 @@
 #include "graph.h"
 #include "digraph.h"
 
-Graph edge_array_to_graph(raw_edge_array&, double);
-Digraph edge_array_to_digraph(raw_edge_array&, double);
+Graph edge_array_to_graph(raw_edge_array&, float);
+Digraph edge_array_to_digraph(raw_edge_array&, float);
 
 #endif //KAPPA_EDGE_ARRAY_TO_GRAPH_H
