+++ src/preprocessor/read_from_disk/tsv_to_graph.cpp
@@ -5,7 +5,7 @@
 #include "tsv_to_graph.h"
 #include "digraph.h"
 
-Digraph tsv_to_digraph(const std::string &tsv_file, char separator, unsigned int v_num, double init_state) {
+Digraph tsv_to_digraph(const std::string &tsv_file, char separator, unsigned int v_num, float init_state) {
     Digraph digraph = Digraph(v_num, init_state);
 
     std::fstream fs;
